en: 移动获取合辑内挑战到collection
