+++ internal/app/api/v1/quest.go
@@ -78,18 +78,3 @@ func UpdateRecommend(c *gin.Context) {
 		Ok(c)
 	}
 }
-
-// GetCollectionQuest 获取合辑内挑战
-func GetCollectionQuest(c *gin.Context) {
-	var r request.GetCollectionQuestRequest
-	_ = c.ShouldBindQuery(&r)
-	r.Address = c.GetString("address")
-	if list, collection, err := srv.GetCollectionQuest(r); err != nil {
-		FailWithMessage(GetMessage(c, "FetchFailed"), c)
-	} else {
-		OkWithDetailed(response.GetCollectionQuestPageResult{
-			List:       list,
-			Collection: collection,
-		}, GetMessage(c, "FetchSuccess"), c)
-	}
-}
