+++ internal/app/dao/quest.go
@@ -232,27 +232,3 @@ func (d *Dao) GetQuestChallengeUserByUUID(uuid string) (res response.GetQuestCha
 func (d *Dao) UpdateQuest(req *model.Quest) (err error) {
 	return d.db.Where("token_id", req.TokenId).Updates(&req).Error
 }
-
-func (d *Dao) GetCollectionQuest(r request.GetCollectionQuestRequest) (questList []response.GetQuestListRes, collection model.Collection, err error) {
-	// 查询合辑信息
-	err = d.db.Model(&model.Collection{}).Where("id", r.ID).First(&collection).Error
-	if err != nil {
-		return questList, collection, err
-	}
-	// 查询合辑内挑战
-	db := d.db.Model(&model.CollectionRelate{}).Joins("left join quest ON collection_relate.quest_id=quest.id").
-		Where("collection_relate.collection_id = ? AND quest.status=1", r.ID)
-	if r.Address != "" {
-		db.Select("quest.*,c.claimed")
-		db.Joins("LEFT JOIN user_challenges c ON quest.token_id = c.token_id AND c.address = ?", r.Address)
-	} else {
-		db.Select("*")
-	}
-	err = db.Order("collection_relate.sort desc").Find(&questList).Error
-	return
-}
-
-func (d *Dao) GetCollectionByID(id int) (collection model.Collection, err error) {
-	err = d.db.Model(&model.Collection{}).Where("id", id).First(&collection).Error
-	return
-}
