+++ internal/app/router/quest.go
@@ -13,7 +13,6 @@ func InitQuestRouter(Router *gin.RouterGroup) {
 		questRouter.GET("", v1.GetQuestList)
 		questRouter.GET("/:id", v1.GetQuest)
 		questRouter.GET("/:id/challengeUsers", v1.GetQuestChallengeUser)
-		questRouter.GET("/collection", v1.GetCollectionQuest)
 	}
 	{
 		questRouterAuth.POST("", v1.AddQuest)
