+++ internal/app/service/quest.go
@@ -117,8 +117,3 @@ func (s *Service) UpdateRecommend(address string, modify request.UpdateRecommend
 	}
 	return nil
 }
-
-// GetCollectionQuest
-func (s *Service) GetCollectionQuest(r request.GetCollectionQuestRequest) (res []response.GetQuestListRes, collection model.Collection, err error) {
-	return s.dao.GetCollectionQuest(r)
-}
